WorkLens/requirements.txt
openai>=2.2.0
mss
pillow
pydantic
win10toast
python-dotenv
ttkbootstrap
